Fixed final circular dependency
